Re-factoring one method in Blob Manager to abstract generation of shared access tokens
