+++ adls-to-blob-cp/src/main/scala/com/starbucks/analytics/blob/BlobManager.scala
@@ -1,8 +1,11 @@
 package com.starbucks.analytics.blob
 
-import com.microsoft.azure.storage.{ CloudStorageAccount, OperationContext }
+import java.util
+
+import com.microsoft.azure.storage.{CloudStorageAccount, OperationContext}
 import com.microsoft.azure.storage.blob._
 import com.typesafe.scalalogging.Logger
+import org.joda.time.{DateTime, DateTimeZone}
 
 import scala.util.Try
 
@@ -49,19 +52,19 @@ object BlobManager {
     containerName:  String,
     f:              (CloudBlobContainer) => R
   ): Try[R] = {
-    def fn(serviceClient: CloudBlobClient): R = {
-      val container = serviceClient.getContainerReference(containerName)
-      val blobRequestOptions = new BlobRequestOptions()
-      val operationContext = new OperationContext()
-      blobRequestOptions.setConcurrentRequestCount(100)
-      operationContext.setLoggingEnabled(true)
-      container.createIfNotExists(
-        BlobContainerPublicAccessType.OFF,
-        blobRequestOptions,
-        operationContext
-      )
-      f(container)
-    }
+      def fn(serviceClient: CloudBlobClient): R = {
+        val container = serviceClient.getContainerReference(containerName)
+        val blobRequestOptions = new BlobRequestOptions()
+        val operationContext = new OperationContext()
+        blobRequestOptions.setConcurrentRequestCount(100)
+        operationContext.setLoggingEnabled(true)
+        container.createIfNotExists(
+          BlobContainerPublicAccessType.OFF,
+          blobRequestOptions,
+          operationContext
+        )
+        f(container)
+      }
     withAzureBlobStoreClient(
       connectionInfo,
       fn
@@ -82,13 +85,46 @@ object BlobManager {
     containerName:  String,
     blobName:       String
   ): Try[CloudBlockBlob] = {
-    def fn(container: CloudBlobContainer): CloudBlockBlob = {
-      container.getBlockBlobReference(blobName)
-    }
+      def fn(container: CloudBlobContainer): CloudBlockBlob = {
+        container.getBlockBlobReference(blobName)
+      }
     withAzureBlobContainer[CloudBlockBlob](
       connectionInfo,
       containerName,
       fn
     )
+  }
+
+  /**
+   * Gets a shared access signature token for the block blob specified
+   *
+   * @param blockBlobReference Reference to the block blob
+   * @param tokenExpirationInMinutes Token expiration time in minutes
+   * @return Token
+   */
+  def getSharedAccessSignatureToken(
+    blockBlobReference:       CloudBlockBlob,
+    tokenExpirationInMinutes: Int
+  ): String = {
+    // Set the expiry time and permissions for the blob.
+    // In this case, the start time is specified as a few minutes in the past, to mitigate clock skew.
+    // The shared access signature will be valid immediately.
+    val sasConstraints = new SharedAccessBlobPolicy()
+    sasConstraints.setSharedAccessStartTime(
+      new DateTime(DateTimeZone.UTC)
+      .plusMinutes(-5)
+      .toDate
+    )
+    sasConstraints.setSharedAccessExpiryTime(
+      new DateTime(DateTimeZone.UTC)
+      .plusMinutes(tokenExpirationInMinutes)
+      .toDate
+    )
+    sasConstraints.setPermissions(util.EnumSet.of(SharedAccessBlobPermissions.READ))
+    val sasToken: String = blockBlobReference.generateSharedAccessSignature(
+      sasConstraints,
+      null
+    )
+    sasToken
   }
 }
